strimzi-0.22.1/python-producer/requirements.txt
python-dotenv==0.16.0
kafka-python==2.0.2